Add files via upload
Behavior info.
There are a few ways to kick a penalty, in this game there is a Deceptive, Power, and Placed kick. Deception will represent when a kicker attempts to fool the goalkeeper with his eyes, by looking at a certain spot and kicking in another direction. A power kick is powerful kick in a certain direction. A placed kick is a carefully placed kick on a specific spot on the goal. This differs from the power kick because it is more accurate. A power kick will make the ball travel faster, but will be less accurate.